Move `processEntities` function to `utils/entities`
